Add a test JSON specification file
This "playground" file is used to nail down what the input should be
like, when we start supporting actual meta-data for symbols.

The main drive behind it is to get a rough idea of what the input file
could look like, with the goal in mind to keep the previous API valid as
much as possible (in other words, the added meta-data support doesn't
drastically change the main logic of how the input is handled, it's just
a layer on top of the existing stuff we have).