Use the new API for mvp1's Derivations and Parser
This one is a bit of a big-bang.

The DerivationSystem, the Main, the Parser, have all been modified to
use the new input system and the new API that supports meta-data etc..

It's a bit of an ego boost to see that my code wasn't half-bad; the
logic stayed identical, and the only changes required were to feed a
different input file with the new schema (obvsly) and change some
function signatures here and there.

For example, the output is now not an ArrayList<String>, but an
ArrayList<Symbol>, with the Symbol class encapsulating a bunch of cool
info, including the String stuff.
Other changes are similar in nature.