Refactor Size and Position into a single Class
Turns out, these two have the exact same role, so they're now one.
Note that this Class represents data in String, not ints/numbers.
That's bcs there is support for both absolute values (ie: a regular
size/position), but also values relative to some LHS producer symbol in
a rule; so there's a need to represent in the JSON data exactly what it
is that you're relative to: the data has to be a String (it's flexible
enough that way without getting messy).