Modify data representation of Deltas
By messing around in the playground, I relaized that the current system
doesn't allow arbitrary, or even some simple computations of position
and size.
The goal is to change the deltas to be ArrayList of Strings; that way,
one can represent any arbitrary number of modifications/deltas: just add
a member to the list.
That is why the delta_pos and delta_size are different (in both the
input file and the Symbol class).