Modify the whole system of data representation and value passing
Quite a big commit (should make them smaller tbf :p)
In esssence, the way deltas work and are applied is much cleaner now.
There is one Position(x, y, z), and one array of Delats to be applied to
this position. This gives us a lot more leeway on the deltas we want to
apply.
Note that the Position can only handle absolute values, not references
like "x", "sz", etc.. Those would go in the DeltaPosition ArrayList.