Apply important HotFix
It was unexpected.
I thought the DerivationSystem worked fined.
At one point, I decided to see what would happen if one used a
left-recursive grammar (with an iter limit ofc).

Then I found a bug; it wasn't in the logic, but the implementation.
I didn't think hard enough about Java's memory system: basically, in the
DerivationSystem I play around with ArrayLists and copy their contents
around. I thought these copies were deep (as in, "real" copies) so
that's how I built the logic, but they were shallow:
the only thing that was copied was the value of the references to the objects,
so that modifying the original would actually affect the copy too.

Fixed it with an ugly-ish hack (serialize to JSON, deserialize from JSON
to a new Object).

Lesson learned !