Implement randomization
It's a bit crude, but it's very easy to understand and to implement.
It's not exactly the way I wanted it, but still is something.

Each Symbol has a rule.
Each rule has a bunch of Symbol.
Each one of these Symbols has a probability associated to it.
That's it.

For future reference, what I had wanted to do, was the following:
Each Symbol has a set of associated rules (as opposed to only one).
It's the rules that have a proba associated to them, not the Symbols.
To modify the system, you would need to change the rules,
crudely as follows: instead of HM<String, Results>, you would have
something like HM<String, Set<Results>>.
I guess that's for future improvements.