Refactor Normal and Delta coordniates into different classes
Because this was an MVP, the presence of unclean code (like say one
class with different purposes) was tolerated.
Only refactoring when really needed, which is now.

Basically:
- Thematically, it's clear that delta coordinates and normal coordinates
fulfill different roles, and you could see this from the confusion you
got when implementing coordinates-related stuff.
For example, a typical occurence (actually, it's even more than that,
it's a property of the system itself) is that the normal coordinates
would only have their x,y,z fields instantiated, and the delta
coordniates would only have their delta fields instantiated.
- This will also help for later, when repurposing the Coordinates class and adding
more stuff to it