Add Symbol::exclusiveDerivation field
This boolean is useful to differentiate between two possible systems of
randomisation one might want to use.
It specifies whether the rule coming from this symbol is exclusive or
not.
An exclusive rule is one where: out of all the RHS symbols of the rule,
only one, and exactly one is chosen.
A non-exclusive rule is one where: every RHS symbol can be chosen or
not, independently. (This is the current default btw).

This future concept of exclusive rules is how I'd get out of the previous
bind I described some time ago: I want a given rule to be able to give
me one of some possible derivations, but a Map cannot have duplicate
keys, and I cannot change the format of the rules (bcs then it'd get too
crazy complex and I want something cleaner).
It's also cool bcs it doesn't bind me to one wy of representing
randomisation: I can just choose whatever I feel is cool atm.