PlayGround
I just played around the playground and the new options implemented.
The input file (playground.json) contains an interesting
DerivationSystem, that uses all the stuff I've implemented until now.
It works as follows:
- There's an initial symbol (called "INITIAL", what a creative name :p),
which produces three symbols "A" next to each other.
BTW: This rule is non-exclusive; and all the A's probas are set to 100,
so they all get created.
- Then Symbol A has a rule associated to it that either creates an "A1",
"A2", or "A3" where it was.
That is an example of an exclusive rule: all the A1/A2/A3's probas are
set to 100 (so they're all as likely to be produced as each other), but
only one of them will actually be produced (and note that here, proba
means something different, it's not a value /100, but a weight; which
makes me wonder whether there's some potential refactoring to be applied
there in the future).