Refactor Coordinate-related functionality
A lot of functions used for modifying Coordinates were lying around in
some classes, but didn't truly belong in any of them;
This refactor groups these "utilities" together in a common Class.

That is an improvement, but I do wonder whether there is some better
solution for that. Until I found such a thing, this is good enough.