Add Symbol:resizeCoefficient field
This triple represents the possible range of the modifications that can
be applied to a given size axis x, y, or z.
The modification happens as x = x + delta. The triple tells us what is the
range of values that delta can take.
So a coef of 1 for x signifies that the modification will look like:
x + delta in [-x/2; +x/2]