Implement Symbol::applyRandomResizeField() method
The function calls applyResizeField(), and passes a cetain delta to it
as a parameter. Then applyResizeField() simply applies the delta to the
given field.

The "randomization" of a resizing is in how this delta is defined.
It works as follows:
- The delta must belong to some interval/range, i.e. it's within this
range that we generate our random nb.
- The range is defined with respect to the appropriate size field, and
its appropriate resizing coefficient.
- We define that range as [- sizeField * 1/2 * coef ; + sizeField * 1/2 * coef]