Add DerivationSystem::materials field
This field wil help in making the rule systems created by users much smaller,
and much less repetitive.
The basic idea is to have a list of commonly used materials, a "theme",
and instead of redefining the same sets of symbols again and again, one
would need to only define it once in this list and then reference it as needed.