Fix the bug isolated previously
The bug was the following:
SOMESYMBOL -> A, A

Then both these A's would have some fields (namely, the deltas) that
would actually refer to the same objects BTS.
Then, when actually calculating positions, there was a side-effect:
the delta field would be modified, then queried.
What this would mean is that the modification would happen when the
first A was processed, then when the time came to the second A, its
internal fields weren't clean anymore.

Next up on the things to do:
Brainstorm. Enforce contracts on the objects tha are supposedly immutable, so that
well, they're actually immutable.
Do so in a defensive way: two things: (1) make stuff private/final and
without setters, and (2) instead of returning the object, return a copy
(a deep one).