Create Symbol constructor from a Symbol and Two Coordinates
This would be used in the DerivationSystem as follows:
One would first calculate what the new size and positon should be;
Then would pass this one to the constructor
along with the relatd Symbol to fill in the rest of the fields.
So then, a new Symbol with the new Coordinates would be returned.