Re-organize input files
The way inputs would be handled is as follows.
There would be an input folder, with a single json file at its root, and
another folder inside, sub.
The file at the root would contain the main derivation system (the one
that defines the initial axiom, etc..)
The files in the sub folder would define additional rules and materials.

That way, one could build more and more complicated systems easily, in a very modular fashion.
For example, a natural way for organizing the files could be thematically:
Each file would hold the definition of say, a building block of your city/thingy/whatever.
Or one could organize files in a scale-like fashion:
All the high level rules for where to place ditricts, neighbourhoods, etc.., in a file.
Then all the rules for how to build a house in another file,
then all the rules for how to decorate said house in another.
These are just examples to illustrate how easy it'd be to build quite elaborate systems.