Implement Parser::getFinalDerivationSystem()
We get the DerivationSystem as follows.
First the "main" system (where the most important stuff is defined) is de-serialized.
Then the content of all the auxiliary/helper systems (those that define
additional stuff, like more non-terminals and more rules) is aggregated into one object.
Finally, all these new rules and symbols are added to the main system,
and that one final sysgtem is the one thagt's processed by the algorithm.