Add Symbol deltaSize reference field
Add the field, getter, and skeleton of getter from ref.
The logic is exactly like that of the Material reference;
You got a String that refers to the ID of a predefined d_size;
then when this symbol is processed a d_size is added to it wrt this ref.