Implement support for DeltaPosition references
Nothing special, apart from some notes for later.

Observe the almost identical logic of deltaPosition, deltaSize, and
Material references.
Note for later to refactor this.

Also another codebase-health kind of thing for later:
Is there a cleaner way to implement your constructors ?
That's where i spent most of my time debugging when implementing the
deltaSize reference stuff; I had forgotten to add it in the constructor
and spent a lot of time figuring out what was wrong. Seems like a
stupid, easily avoidable mistake, I don't like seeing a function
with 10 parameters and this is why :p