Re-haul the input/output system
This would probably be better if split up in several smaller commits,
but at least all the changes have one common theme.
This commit allows one to specify pretty much everythung IO related:
where to take the input from, where to write the output,
and even the names of the main input file and of the auxiliary subfolder.
If some of these are not provided by the command-line, a default is used
(`input/`, `output/`, `playground.json`, `sub/` respectively).