Add CLI Interrogation Phase
The command line input is now completely revamped.
Instead of the old ordered non-named arguments,
the app now supports unordered, named, non-required arguments.

Supplying arguments is now so so much more natural.

I didn't want to use a library for this at first, bcs I wanted to do it
myself as a learning experience (and that was fine for my initial purposes),
but using a library for input parsing and general CLI stuff makes much more sense now.