Cleanup
Removed everything that had to do with rotations.
It was clunky, and i realized that with the current system, you can
already create rules with random rotation, so having a Rotation class is
actually useless (and possibly harmful :p)

You'd just need to do something like this:
symbol A of size sx, sy, sz
    with an exclusive rule
    => creates:
        - B of size sx, sy, sz
        - B of size sy, sx, sz
        - etc..
With the current system, you can give weights to each rotation, so you
can even decide how likely each one is.