Add extensive validation of Symbol state
Added pretty much everything that needs to be checked in a symbol's state.
For now, the only symbols checked are those in the ruleset.

Later on, I'll add checks for the symbols in the axiom.
Once both all the symbols in the axiom and all the symbols in all the
rules are checked, it can be showned (by construction) that any resulting
symbol at any derivation step will also be valid.

In other words, there's no need to check every symbol produced;
you just need to check that the initial state is valid and that any transition also is,
from which you can deduce that any intermediate step is also valid.