Add metadata field to Symbol class
I want to make this more general, and allow for greater support of
non-critical metadata.
This HM would simply hold this non-critical stuff, and make for much
less duplication.
Also, it would allow for the rest of what I actually want to do: make
this generator about **just** generating a CFG string, and have
additional scripts on the side actually take care of the domain specific translations.