Remove materials from the generator
Remove everything and anything related to materials.
(The Symbol:material/material_ref fields, Material class, etc..)

The reason is that the material is not essential to the derivation of the symbols.
It's only used in one case (with minecraft).

This is one of the complexities the metadata map removes;
Now, if one wants to use materials, just add the necessary fields to the
required symbols, then define some global materials in DS::metas, and that's it.