Basic skeleton of the minecraft interpretation
The basic flow of this tool will be as follows:
- Make the rules.
- Feed them to the Kukkura generator.
- Pipe its output to various interpreters
(a Minecraft interpreter, a Terasology one, ect..).
- Use the output of the interpreter you used
for whatever purpose you'd be interested in.

So for example here, minecraft.py will be the script to use for the
minecraft interpreter, etc..
All the others would also be in this folder.