Add skeleton of a BadSymbol exception class
Different interpreters need different fields for the symbols they handle.
This cannot be validated from Kukkura's side:
Kukkura's output is syntatically valid, but it cannot check whether said
output is semantically valid with respect to specific interpretations.
This is why the validation of meta-data has to be made separately
by each interpreter.